pyproject.toml
[project]
name = "mapcident"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "django-ninja>=1.5.3",
    "django-ninja-extra>=0.31.0",
    "django-ninja-jwt>=5.4.4",
    "django-storages>=1.14.6",
    "email-validator>=2.3.0",
    "gunicorn>=25.1.0",
    "h3>=4.4.2",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
]
